(MOT-4107) refactor(harness): centralize integration runtime ownership (#533)
* (MOT-4107) refactor(harness): centralize integration runtime ownership

* (MOT-4107) refactor(harness): decompose integration scenario runner (#534)
