+++ harness/evals/integration/src/grader.rs
@@ -1,14 +1,22 @@
-//! Pure-code grading (spec § Oracles): every invariant reads collected
-//! evidence only — no engine calls, no subject mutation. Invariant ids and
-//! typed invariant dispatch and compiler-side constructors live in the
-//! scenario contract registry.
+//! Pure-code grading: every invariant reads collected evidence only, with no
+//! engine calls or subject mutation.
 
-use serde_json::{json, Map, Value};
+mod dispatch;
+mod helpers;
+mod router_target;
+mod send_transcript;
+mod stabilize;
+mod status_lifecycle;
 
-use crate::types::recorder::{RecorderEventKind, RecorderEventV1};
-use crate::types::scenario::{
-    DecodedInvariantV1, InvariantDecodeError, InvariantResultV1, InvariantSpecV1,
-};
+#[cfg(test)]
+mod tests;
+
+use serde_json::Value;
+
+use crate::types::recorder::RecorderEventV1;
+use crate::types::scenario::{InvariantResultV1, InvariantSpecV1};
+
+type GradeOutcome = (bool, Value, Value, Vec<&'static str>);
 
 /// Everything the grader may look at, exactly as persisted to the artifact
 /// directory (grading twice over the same evidence is byte-deterministic).
@@ -30,9 +38,9 @@ pub fn grade(specs: &[InvariantSpecV1], evidence: &Evidence) -> Vec<InvariantRes
     specs
         .iter()
         .map(|spec| {
-            let (passed, mut expected, mut actual, refs) = grade_one(spec, evidence);
-            stabilize_value(&mut expected, evidence);
-            stabilize_value(&mut actual, evidence);
+            let (passed, mut expected, mut actual, refs) = dispatch::grade_one(spec, evidence);
+            stabilize::stabilize_value(&mut expected, evidence);
+            stabilize::stabilize_value(&mut actual, evidence);
             InvariantResultV1 {
                 id: spec.id.clone(),
                 passed,
@@ -42,569 +50,4 @@ pub fn grade(specs: &[InvariantSpecV1], evidence: &Evidence) -> Vec<InvariantRes
             }
         })
         .collect()
-}
-
-fn grade_one(
-    spec: &InvariantSpecV1,
-    evidence: &Evidence,
-) -> (bool, Value, Value, Vec<&'static str>) {
-    let decoded = match spec.decode() {
-        Ok(decoded) => decoded,
-        Err(InvariantDecodeError::UnknownId(id)) => {
-            return (
-                false,
-                json!({ "error": "unknown invariant id" }),
-                json!({ "id": id }),
-                vec![],
-            );
-        }
-    };
-    match decoded {
-        DecodedInvariantV1::SendFlags(params) => {
-            let expected = json!({
-                "accepted": true,
-                "merged": flag(params, "merged"),
-                "queued": flag(params, "queued"),
-                "deduplicated": flag(params, "deduplicated"),
-            });
-            // Absent optional flags normalize to false (spec § send contract).
-            let actual = match &evidence.send_response {
-                Some(resp) => json!({
-                    "accepted": resp.get("accepted") == Some(&Value::Bool(true)),
-                    "merged": resp.get("merged") == Some(&Value::Bool(true)),
-                    "queued": resp.get("queued") == Some(&Value::Bool(true)),
-                    "deduplicated": resp.get("deduplicated") == Some(&Value::Bool(true)),
-                }),
-                None => Value::Null,
-            };
-            (
-                actual == expected,
-                expected,
-                actual,
-                vec!["send-response.json"],
-            )
-        }
-
-        DecodedInvariantV1::TranscriptMessageCounts(params) => {
-            let mut counts = std::collections::BTreeMap::new();
-            for role in ["user", "assistant", "function_result"] {
-                counts.insert(role.to_string(), 0u64);
-            }
-            for item in &evidence.transcript {
-                if let Some(role) = item
-                    .get("message")
-                    .and_then(|m| m.get("role"))
-                    .and_then(Value::as_str)
-                {
-                    *counts.entry(role.to_string()).or_insert(0) += 1;
-                }
-            }
-            let expected = Value::Object(params.clone());
-            let actual = serde_json::to_value(&counts).expect("counts serialize");
-            let passed = params.iter().all(|(role, want)| {
-                counts.get(role).copied().unwrap_or(0) == want.as_u64().unwrap_or(u64::MAX)
-            });
-            (passed, expected, actual, vec!["transcript.json"])
-        }
-
-        DecodedInvariantV1::TranscriptAssistantText(params) => {
-            let expected = Value::Object(params.clone());
-            let last_assistant = evidence
-                .transcript
-                .iter()
-                .filter_map(|item| item.get("message"))
-                .rfind(|m| m.get("role").and_then(Value::as_str) == Some("assistant"));
-            let actual = match last_assistant {
-                Some(message) => {
-                    let text: String = message
-                        .get("content")
-                        .and_then(Value::as_array)
-                        .map(|blocks| {
-                            blocks
-                                .iter()
-                                .filter(|b| b.get("type").and_then(Value::as_str) == Some("text"))
-                                .filter_map(|b| b.get("text").and_then(Value::as_str))
-                                .collect()
-                        })
-                        .unwrap_or_default();
-                    json!({ "text": text, "usage": message.get("usage").cloned().unwrap_or(Value::Null) })
-                }
-                None => Value::Null,
-            };
-            let passed = match &actual {
-                Value::Null => false,
-                actual => {
-                    actual.get("text") == params.get("text")
-                        && params.get("usage").is_none_or(|want| {
-                            literal_subset(want, actual.get("usage").unwrap_or(&Value::Null))
-                        })
-                }
-            };
-            (passed, expected, actual, vec!["transcript.json"])
-        }
-
-        DecodedInvariantV1::TranscriptNoDuplicates(_) => {
-            let mut seen = std::collections::BTreeSet::new();
-            let mut duplicates = Vec::new();
-            for item in &evidence.transcript {
-                if let Some(entry_id) = item.get("entry_id").and_then(Value::as_str) {
-                    if !seen.insert(entry_id.to_string()) {
-                        duplicates.push(entry_id.to_string());
-                    }
-                }
-            }
-            let passed = duplicates.is_empty();
-            (
-                passed,
-                json!({ "duplicate_entry_ids": [] }),
-                json!({ "duplicate_entry_ids": duplicates }),
-                vec!["transcript.json"],
-            )
-        }
-
-        // Every dispatched `function_call` must be closed by a durable
-        // `function_result` referencing its id — an interrupted call left
-        // dangling poisons every later provider request on the session
-        // (iii-hq/workers#507).
-        DecodedInvariantV1::TranscriptCallsClosed(_) => {
-            let mut call_ids = Vec::new();
-            let mut result_ids = std::collections::BTreeSet::new();
-            for item in &evidence.transcript {
-                let Some(message) = item.get("message") else {
-                    continue;
-                };
-                match message.get("role").and_then(Value::as_str) {
-                    Some("assistant") => {
-                        for block in message
-                            .get("content")
-                            .and_then(Value::as_array)
-                            .into_iter()
-                            .flatten()
-                        {
-                            if block.get("type").and_then(Value::as_str) == Some("function_call") {
-                                if let Some(id) = block.get("id").and_then(Value::as_str) {
-                                    call_ids.push(id.to_string());
-                                }
-                            }
-                        }
-                    }
-                    Some("function_result") => {
-                        if let Some(id) = message.get("function_call_id").and_then(Value::as_str) {
-                            result_ids.insert(id.to_string());
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            let dangling: Vec<&String> = call_ids
-                .iter()
-                .filter(|id| !result_ids.contains(*id))
-                .collect();
-            let passed = dangling.is_empty();
-            (
-                passed,
-                json!({ "dangling_function_calls": [] }),
-                json!({
-                    "dangling_function_calls": dangling,
-                    "calls": call_ids,
-                    "results": result_ids,
-                }),
-                vec!["transcript.json"],
-            )
-        }
-
-        DecodedInvariantV1::TranscriptFunctionResult(params) => {
-            let expected = Value::Object(params.clone());
-            let expected_call_id = params
-                .get("function_call_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default();
-            let results: Vec<&Value> = evidence
-                .transcript
-                .iter()
-                .filter_map(|item| item.get("message"))
-                .filter(|m| m.get("role").and_then(Value::as_str) == Some("function_result"))
-                .filter(|m| {
-                    m.get("function_call_id").and_then(Value::as_str) == Some(expected_call_id)
-                })
-                .collect();
-            let actual = json!(results);
-            let passed = results.len() == 1 && literal_subset(&expected, results[0]);
-            (passed, expected, actual, vec!["transcript.json"])
-        }
-
-        DecodedInvariantV1::StatusTerminal(params) => {
-            let expected = Value::Object(params.clone());
-            let actual = json!({
-                "status": evidence.status.get("status").cloned().unwrap_or(Value::Null),
-                "pending_calls": evidence
-                    .status
-                    .get("pending_function_calls")
-                    .and_then(Value::as_array)
-                    .map(Vec::len)
-                    .unwrap_or(0),
-                "queued_messages": evidence
-                    .status
-                    .get("queued")
-                    .and_then(Value::as_array)
-                    .map(Vec::len)
-                    .unwrap_or(0),
-                "children": evidence
-                    .status
-                    .get("children")
-                    .and_then(Value::as_array)
-                    .map(Vec::len)
-                    .unwrap_or(0),
-            });
-            let passed = actual.get("status") == params.get("status")
-                && actual.get("pending_calls") == params.get("pending_calls")
-                && actual.get("queued_messages") == params.get("queued_messages")
-                && actual.get("children").and_then(Value::as_u64).unwrap_or(0) == 0;
-            (passed, expected, actual, vec!["status.json"])
-        }
-
-        DecodedInvariantV1::LifecycleCompletedOnce(params) => {
-            let allow_identical_duplicates = params
-                .get("allow_identical_duplicates")
-                .and_then(Value::as_bool)
-                .unwrap_or(true);
-            let expected = json!({
-                "delivery_count_valid": true,
-                "all_deliveries_identical_after_timestamp_normalization": true,
-                "session_and_turn_match": true,
-            });
-            let payloads: Vec<Value> = evidence
-                .recorder_events
-                .iter()
-                .filter(|e| e.kind == RecorderEventKind::Lifecycle)
-                .map(|e| e.payload.clone())
-                .collect();
-            let normalized: Vec<Value> = payloads
-                .iter()
-                .map(|p| {
-                    let mut copy = p.clone();
-                    if let Some(map) = copy.as_object_mut() {
-                        map.remove("timestamp");
-                    }
-                    copy
-                })
-                .collect();
-            let identical = normalized.windows(2).all(|w| w[0] == w[1]);
-            let bound_ok = payloads.iter().all(|p| {
-                p.get("session_id").and_then(Value::as_str) == Some(evidence.session_id.as_str())
-                    && match &evidence.turn_id {
-                        Some(turn) => p.get("turn_id").and_then(Value::as_str) == Some(turn),
-                        None => false,
-                    }
-            });
-            let delivery_count_valid = if allow_identical_duplicates {
-                !payloads.is_empty()
-            } else {
-                payloads.len() == 1
-            };
-            let actual = json!({
-                "delivery_count_valid": delivery_count_valid,
-                "all_deliveries_identical_after_timestamp_normalization": identical,
-                "session_and_turn_match": bound_ok,
-            });
-            let passed = delivery_count_valid && identical && bound_ok;
-            (passed, expected, actual, vec!["lifecycle-events.json"])
-        }
-
-        DecodedInvariantV1::RouterGenerationsConsumed(params) => {
-            let want = params.get("count").and_then(Value::as_u64);
-            let expected = json!({
-                "count": want,
-                "unused_expectations": 0,
-            });
-            let actual = json!({
-                "count": evidence.generations_consumed,
-                "unused_expectations": evidence.generations_total - evidence.generations_consumed,
-            });
-            let passed = Some(evidence.generations_consumed) == want
-                && evidence.generations_consumed == evidence.generations_total;
-            (passed, expected, actual, vec!["router-calls.json"])
-        }
-
-        DecodedInvariantV1::TargetCalls(params) => {
-            let function_id = params
-                .get("function_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default();
-            let want_count = params.get("count").and_then(Value::as_u64).unwrap_or(0);
-            let calls: Vec<&RecorderEventV1> = evidence
-                .recorder_events
-                .iter()
-                .filter(|e| e.kind == RecorderEventKind::TargetCall && e.function_id == function_id)
-                .collect();
-            let payload_ok = match params.get("payload") {
-                Some(want) => calls.iter().all(|c| &c.payload == want),
-                None => true,
-            };
-            // `payload_subset`: every listed member must match (deep subset)
-            // in every call — used for envelope payloads (hook inputs) whose
-            // full shape carries ids/timestamps.
-            let subset_ok = match params.get("payload_subset") {
-                Some(want) => calls.iter().all(|c| literal_subset(want, &c.payload)),
-                None => true,
-            };
-            let expected = Value::Object(params.clone());
-            let actual = json!({
-                "count": calls.len(),
-                "payloads": calls.iter().map(|c| c.payload.clone()).collect::<Vec<_>>(),
-            });
-            let passed = calls.len() as u64 == want_count && payload_ok && subset_ok;
-            (passed, expected, actual, vec!["target-calls.json"])
-        }
-    }
-}
-
-fn flag(params: &Map<String, Value>, key: &str) -> bool {
-    params.get(key).and_then(Value::as_bool).unwrap_or(false)
-}
-
-fn literal_subset(expected: &Value, actual: &Value) -> bool {
-    crate::matcher::subset_with_array_policy(expected, actual, crate::matcher::ArrayPolicy::Exact)
-        .is_none()
-}
-
-/// Keep result.json stable while detailed evidence retains raw run-scoped
-/// values. Diagnostic artifacts still contain every original id and
-/// timestamp.
-fn stabilize_value(value: &mut Value, evidence: &Evidence) {
-    match value {
-        Value::String(text) => {
-            replace_identity(text, &evidence.run_id, "{{run_id}}");
-            replace_identity(text, &evidence.session_id, "{{session_id}}");
-            if let Some(turn_id) = &evidence.turn_id {
-                replace_identity(text, turn_id, "{{turn_id}}");
-            }
-        }
-        Value::Array(items) => {
-            for item in items {
-                stabilize_value(item, evidence);
-            }
-        }
-        Value::Object(map) => {
-            map.remove("timestamp");
-            map.remove("received_at");
-            for item in map.values_mut() {
-                stabilize_value(item, evidence);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn replace_identity(text: &mut String, identity: &str, placeholder: &str) {
-    if !identity.is_empty() && text.contains(identity) {
-        *text = text.replace(identity, placeholder);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::script::SchemaVersion1;
-
-    fn event(kind: RecorderEventKind, function_id: &str, payload: Value) -> RecorderEventV1 {
-        RecorderEventV1 {
-            schema_version: SchemaVersion1::V1,
-            run_id: "r".into(),
-            sequence: 1,
-            kind,
-            function_id: function_id.into(),
-            payload,
-            received_at: "2026-07-15T00:00:00Z".into(),
-        }
-    }
-
-    fn base_evidence() -> Evidence {
-        Evidence {
-            run_id: "r".into(),
-            session_id: "s_1".into(),
-            turn_id: Some("t_1".into()),
-            send_response: Some(json!({ "session_id": "s_1", "turn_id": "t_1", "accepted": true })),
-            status: json!({ "status": "completed", "pending_function_calls": [], "children": [] }),
-            transcript: vec![],
-            generations_consumed: 1,
-            generations_total: 1,
-            recorder_events: vec![],
-        }
-    }
-
-    fn spec(id: &str, params: Value) -> InvariantSpecV1 {
-        InvariantSpecV1 {
-            id: id.into(),
-            parameters: params.as_object().cloned().unwrap_or_default(),
-        }
-    }
-
-    #[test]
-    fn send_flags_normalize_absent_to_false() {
-        let evidence = base_evidence();
-        let results = grade(
-            &[spec(
-                "send.flags",
-                json!({ "merged": false, "queued": false, "deduplicated": false }),
-            )],
-            &evidence,
-        );
-        assert!(results[0].passed, "{:?}", results[0]);
-    }
-
-    /// Spec-mandated: the duplicate-detection rule is unit-tested with
-    /// synthetic evidence containing two target calls.
-    #[test]
-    fn target_calls_fails_on_duplicate_execution() {
-        let mut evidence = base_evidence();
-        evidence.recorder_events = vec![
-            event(
-                RecorderEventKind::TargetCall,
-                "r::record",
-                json!({ "value": "expected" }),
-            ),
-            event(
-                RecorderEventKind::TargetCall,
-                "r::record",
-                json!({ "value": "expected" }),
-            ),
-        ];
-        let results = grade(
-            &[spec(
-                "target.calls",
-                json!({ "function_id": "r::record", "count": 1, "payload": { "value": "expected" } }),
-            )],
-            &evidence,
-        );
-        assert!(!results[0].passed, "two executions must fail exactly-once");
-        assert_eq!(results[0].actual["count"], 2);
-    }
-
-    #[test]
-    fn lifecycle_accepts_identical_duplicates_and_rejects_conflicts() {
-        let completed = json!({
-            "session_id": "s_1", "turn_id": "t_1", "status": "completed", "timestamp": 1
-        });
-        let mut dup = completed.clone();
-        dup["timestamp"] = json!(99); // identical modulo timestamp
-        let mut evidence = base_evidence();
-        evidence.recorder_events = vec![
-            event(
-                RecorderEventKind::Lifecycle,
-                "integration-recorder::lifecycle",
-                completed.clone(),
-            ),
-            event(
-                RecorderEventKind::Lifecycle,
-                "integration-recorder::lifecycle",
-                dup,
-            ),
-        ];
-        let ok = grade(
-            &[spec(
-                "lifecycle.completed_once",
-                json!({ "allow_identical_duplicates": true }),
-            )],
-            &evidence,
-        );
-        assert!(ok[0].passed, "{:?}", ok[0]);
-
-        let strict = grade(
-            &[spec(
-                "lifecycle.completed_once",
-                json!({ "allow_identical_duplicates": false }),
-            )],
-            &evidence,
-        );
-        assert!(
-            !strict[0].passed,
-            "strict lifecycle expectation must reject duplicate delivery"
-        );
-
-        let mut conflicting = completed;
-        conflicting["status"] = json!("failed");
-        evidence.recorder_events.push(event(
-            RecorderEventKind::Lifecycle,
-            "integration-recorder::lifecycle",
-            conflicting,
-        ));
-        let bad = grade(&[spec("lifecycle.completed_once", json!({}))], &evidence);
-        assert!(!bad[0].passed, "conflicting terminals must fail");
-    }
-
-    #[test]
-    fn function_result_expectations_select_their_call_id() {
-        let mut evidence = base_evidence();
-        evidence.transcript = vec![
-            json!({
-                "message": {
-                    "role": "function_result",
-                    "function_call_id": "call-1",
-                    "function_id": "r::one",
-                    "content": []
-                }
-            }),
-            json!({
-                "message": {
-                    "role": "function_result",
-                    "function_call_id": "call-2",
-                    "function_id": "r::two",
-                    "content": []
-                }
-            }),
-        ];
-        let results = grade(
-            &[
-                spec(
-                    "transcript.function_result",
-                    json!({ "function_call_id": "call-1", "function_id": "r::one" }),
-                ),
-                spec(
-                    "transcript.function_result",
-                    json!({ "function_call_id": "call-2", "function_id": "r::two" }),
-                ),
-            ],
-            &evidence,
-        );
-        assert!(results.iter().all(|result| result.passed), "{results:?}");
-    }
-
-    #[test]
-    fn unused_router_expectation_fails_generations_consumed() {
-        let mut evidence = base_evidence();
-        evidence.generations_total = 2;
-        let results = grade(
-            &[spec("router.generations_consumed", json!({ "count": 1 }))],
-            &evidence,
-        );
-        assert!(!results[0].passed);
-        assert_eq!(results[0].actual["unused_expectations"], 1);
-    }
-
-    #[test]
-    fn unknown_invariant_id_fails_closed() {
-        let results = grade(&[spec("nonsense.id", json!({}))], &base_evidence());
-        assert!(!results[0].passed);
-    }
-
-    #[test]
-    fn known_invariants_preserve_lossless_parameter_maps() {
-        let results = grade(
-            &[
-                spec("send.flags", json!({ "note": "ignored" })),
-                spec("transcript.message_counts", json!({ "assistant": 0 })),
-                spec("target.calls", json!({})),
-                spec("transcript.calls_closed", json!({ "note": "ignored" })),
-                spec("transcript.no_duplicates", json!({ "note": "ignored" })),
-            ],
-            &base_evidence(),
-        );
-        assert!(
-            results.iter().all(|result| result.passed),
-            "missing and extra parameters keep their legacy semantics: {results:?}"
-        );
-        assert_eq!(results[1].expected, json!({ "assistant": 0 }));
-        assert_eq!(results[2].expected, json!({}));
-    }
 }
