+++ harness/evals/integration/src/stack/config.rs
@@ -1,0 +1,100 @@
+use std::path::PathBuf;
+
+use serde_json::{json, Value};
+
+use super::RunLayout;
+
+/// Start order matters: the harness retries `queue::define` only briefly at
+/// boot, so queue precedes harness. The harness itself is spawned after Arm.
+pub const WORKER_START_ORDER: [&str; 4] = [
+    "queue",
+    "iii-directory",
+    "session-manager",
+    "context-manager",
+];
+
+/// No provider keys or developer secrets leak into the subject stack.
+pub(crate) const ENV_ALLOWLIST: [&str; 4] = ["PATH", "HOME", "LANG", "RUST_LOG"];
+
+pub fn render_engine_yaml(layout: &RunLayout, port: u16) -> String {
+    format!(
+        r#"# Generated by harness-integration; one engine per scenario run.
+workers:
+  - name: iii-worker-manager
+    config:
+      host: 127.0.0.1
+      port: {port}
+  - name: configuration
+    config:
+      adapter:
+        name: fs
+        config:
+          directory: {config_dir}
+  - name: iii-state
+  - name: iii-stream
+  - name: iii-cron
+"#,
+        config_dir = layout.configuration_dir().display(),
+    )
+}
+
+/// Seed YAML per configurable worker.
+pub fn render_seed(worker: &str, layout: &RunLayout) -> Option<Value> {
+    match worker {
+        "session-manager" => Some(json!({
+            "adapter": {
+                "name": "fs",
+                "config": { "data_dir": layout.session_data_dir().to_string_lossy() }
+            }
+        })),
+        "context-manager" => Some(json!({
+            "lease_dir": layout.leases_dir().to_string_lossy()
+        })),
+        "queue" => Some(json!({
+            "adapter": {
+                "name": "builtin",
+                "config": {
+                    "store_method": "file_based",
+                    "file_path": layout.queue_path().to_string_lossy(),
+                    "save_interval_ms": 1000
+                }
+            }
+        })),
+        "iii-directory" => Some(json!({
+            "skills_folder": layout.skills_dir().to_string_lossy()
+        })),
+        // Avoid a run-path-dependent line in the expected system prompt.
+        "harness" => Some(json!({
+            "default_filesystem_root": "off"
+        })),
+        _ => None,
+    }
+}
+
+pub fn expected_config_entries(layout: &RunLayout) -> Vec<(String, Value)> {
+    WORKER_START_ORDER
+        .iter()
+        .filter_map(|worker| render_seed(worker, layout).map(|seed| (worker.to_string(), seed)))
+        .collect()
+}
+
+pub fn expected_harness_config_entry(layout: &RunLayout) -> Vec<(String, Value)> {
+    render_seed("harness", layout)
+        .map(|seed| vec![("harness".to_string(), seed)])
+        .unwrap_or_default()
+}
+
+pub(crate) fn write_engine_config(layout: &RunLayout, port: u16) -> anyhow::Result<PathBuf> {
+    let path = layout.engine_config_path();
+    std::fs::write(&path, render_engine_yaml(layout, port))?;
+    Ok(path)
+}
+
+pub(crate) fn write_seed(worker: &str, layout: &RunLayout) -> anyhow::Result<Option<PathBuf>> {
+    let Some(seed) = render_seed(worker, layout) else {
+        return Ok(None);
+    };
+    let path = layout.seed_path(worker)?;
+    std::fs::write(&path, serde_yaml::to_string(&seed)?)?;
+    Ok(Some(path))
+}
